refact(visualizations): extract ResourceHeatmap inline fetching to useResourceMetrics (#6086)
